Changed the precision of the Audio_Signal from float to double, in order to extend the kind of signal that can be entered (i.e. energy, which requires a far greater dynamic range). This eliminated a rounding error we bumped into.

+++ Ana_CenterTime.cs
@@ -62,7 +62,7 @@ namespace PachydermGH
             DA.GetData<int>(1, ref Cx);
 
             List<double> CT = new List<double>();
-            foreach (float[] f in ETC.Value)
+            foreach (double[] f in ETC.Value)
             {
                 double[] s = new double[f.Length];
                 int start = 0;
